Fixed pointer/long int issue with type-cast, seems to be the intent

+++ Planet.h
@@ -7,13 +7,15 @@ class Planet{
 	private:
 	//the "this" pointer should be a pointer that holds the address of the Planet object (this class). 
 		//int id;  //If we are using "this" for the address which is already done in the constructor, then we probably don't need this variable.
+		//long int id = (long int)this;
 		int pos;
 		int distance;
 		char type;
 	public:
 		Planet(int); //Should we also have a no-argument creator as well?
         int orbit();
-		long int getID() {return this;} //This needs to be a long int to hold the address -- apparently the address can be stored in a variable that is not a pointer?
+		long int getID() {return (long int)this;} //This needs to be a long int to hold the address -- apparently the address can be stored in a variable that is not a pointer?
+												  //Stack overflow confirmed that a simple type-casting conversion by putting the desired type in front of the pointer will allow this to happen.
         int getDistance() {return this->distance;}
         int getPos() {return this->pos;}
 		char getType() {return this->type;}
